Default theme is now black. Fixed instrument tests and lint. Fixed canceling in OperationsService. Progress dialogs are now shown according to service stare instead of task lifecycle.

+++ purefm/src/main/java/com/docd/purefm/operations/DeleteOperation.java
@@ -75,6 +75,9 @@ final class DeleteOperation extends Operation<GenericFile, ArrayList<GenericFile
 
             final Shell shell = ShellHolder.getShell();
             for (final GenericFile file : files) {
+                if (isCanceled()) {
+                    break;
+                }
                 final File fileFile = file.toFile();
                 if (CommandLine.execute(shell, new CommandRemove(fileFile))) {
                     filesAffected.add(file);
@@ -89,6 +92,9 @@ final class DeleteOperation extends Operation<GenericFile, ArrayList<GenericFile
 
         } else {
             for (final GenericFile file : files) {
+                if (isCanceled()) {
+                    break;
+                }
                 if (file.delete()) {
                     filesAffected.add(file);
                 } else {
